css-align-3_dev/html/place-items-shorthand-003.htm

<!DOCTYPE html>
<html><head><title>CSS Box Alignment: place-items shorthand - initial value</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez">
<link href="http://www.w3.org/TR/css3-align/#place-items-property" rel="help">
<meta content="Check that place-items's 'initial' value expands to 'align-items' and 'justify-items'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-items: start;
        justify-items: end;
    }
</style>
</head><body><div id="log"></div>
<div id="test"></div>
<script>
    var div = document.getElementById("test");
    div.setAttribute("style", "place-items: initial");
    var style = getComputedStyle(div);

    test(function() {
        assert_equals(style.getPropertyValue("place-items"),
            "normal normal", "place-items resolved value");
    }, "Check place-items: initial - resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("align-items"),
            "normal", "place-items specified value for align-items");
    }, "Check place-items: initial - align-items resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-items"),
            "normal", "place-items specified value for justify-items");
    }, "Check place-items: initial - justify-items resolved value");
</script>
</body></html>
